+++ src/@sdk/index.ts
@@ -16,13 +16,13 @@ import {
   MapFn,
   QueryShape,
   WatchMapFn,
-  WatchQueryData
+  WatchQueryData,
 } from "./types";
 import {
   getErrorsFromData,
   getMappedData,
   isDataEmpty,
-  mergeEdges
+  mergeEdges,
 } from "./utils";
 
 import { UserDetails } from "./queries/types/UserDetails";
@@ -88,6 +88,11 @@ export class SaleorAPI {
     data.me ? data.me.checkout : null
   );
 
+  getVariantsProducts = this.watchQuery(
+    QUERIES.VariantsProducts,
+    data => data.productVariants
+  );
+
   setUserDefaultAddress = this.fireQuery(
     MUTATIONS.AddressTypeUpdate,
     data => data!.accountSetDefaultAddress
@@ -101,6 +106,16 @@ export class SaleorAPI {
   setCheckoutShippingAddress = this.fireQuery(
     MUTATIONS.UpdateCheckoutShippingAddress,
     data => data!.checkoutShippingAddressUpdate
+  );
+
+  setAddCheckoutPromoCode = this.fireQuery(
+    MUTATIONS.AddCheckoutPromoCode,
+    data => data!.checkoutAddPromoCode
+  );
+
+  setRemoveCheckoutPromoCode = this.fireQuery(
+    MUTATIONS.RemoveCheckoutPromoCode,
+    data => data!.checkoutRemovePromoCode
   );
 
   setDeleteUserAddress = this.fireQuery(
@@ -129,6 +144,8 @@ export class SaleorAPI {
   );
 
   setPasswordChange = this.fireQuery(MUTATIONS.PasswordChange, data => data);
+
+  setPassword = this.fireQuery(MUTATIONS.SetPassword, data => data);
 
   private client: ApolloClient<any>;