+++ src/@sdk/mutations/types/CreateCheckout.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { CheckoutCreateInput } from "./../../types/globalTypes";
+import { CheckoutCreateInput } from "./../../../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: CreateCheckout
@@ -43,11 +43,6 @@ export interface CreateCheckout_checkoutCreate_checkout_availablePaymentGateways
    * Payment gateway client configuration.
    */
   config: CreateCheckout_checkoutCreate_checkout_availablePaymentGateways_config[];
-}
-
-export interface CreateCheckout_checkoutCreate_checkout_user {
-  __typename: "User";
-  email: string;
 }
 
 export interface CreateCheckout_checkoutCreate_checkout_totalPrice_gross {
@@ -169,14 +164,6 @@ export interface CreateCheckout_checkoutCreate_checkout_billingAddress {
   country: CreateCheckout_checkoutCreate_checkout_billingAddress_country;
   countryArea: string;
   phone: string | null;
-  /**
-   * Address is user's default billing address.
-   */
-  isDefaultBillingAddress: boolean | null;
-  /**
-   * Address is user's default shipping address.
-   */
-  isDefaultShippingAddress: boolean | null;
 }
 
 export interface CreateCheckout_checkoutCreate_checkout_shippingAddress_country {
@@ -210,14 +197,6 @@ export interface CreateCheckout_checkoutCreate_checkout_shippingAddress {
   country: CreateCheckout_checkoutCreate_checkout_shippingAddress_country;
   countryArea: string;
   phone: string | null;
-  /**
-   * Address is user's default billing address.
-   */
-  isDefaultBillingAddress: boolean | null;
-  /**
-   * Address is user's default shipping address.
-   */
-  isDefaultShippingAddress: boolean | null;
 }
 
 export interface CreateCheckout_checkoutCreate_checkout_availableShippingMethods_price {
@@ -533,6 +512,18 @@ export interface CreateCheckout_checkoutCreate_checkout_lines {
   variant: CreateCheckout_checkoutCreate_checkout_lines_variant;
 }
 
+export interface CreateCheckout_checkoutCreate_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface CreateCheckout_checkoutCreate_checkout {
   __typename: "Checkout";
   /**
@@ -544,7 +535,6 @@ export interface CreateCheckout_checkoutCreate_checkout {
    * The ID of the object.
    */
   id: string;
-  user: CreateCheckout_checkoutCreate_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -572,6 +562,14 @@ export interface CreateCheckout_checkoutCreate_checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (CreateCheckout_checkoutCreate_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: CreateCheckout_checkoutCreate_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface CreateCheckout_checkoutCreate {