+++ src/@sdk/queries/types/UserCheckoutDetails.ts
@@ -30,11 +30,6 @@ export interface UserCheckoutDetails_me_checkout_availablePaymentGateways {
   config: UserCheckoutDetails_me_checkout_availablePaymentGateways_config[];
 }
 
-export interface UserCheckoutDetails_me_checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface UserCheckoutDetails_me_checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -154,14 +149,6 @@ export interface UserCheckoutDetails_me_checkout_billingAddress {
   country: UserCheckoutDetails_me_checkout_billingAddress_country;
   countryArea: string;
   phone: string | null;
-  /**
-   * Address is user's default billing address.
-   */
-  isDefaultBillingAddress: boolean | null;
-  /**
-   * Address is user's default shipping address.
-   */
-  isDefaultShippingAddress: boolean | null;
 }
 
 export interface UserCheckoutDetails_me_checkout_shippingAddress_country {
@@ -195,14 +182,6 @@ export interface UserCheckoutDetails_me_checkout_shippingAddress {
   country: UserCheckoutDetails_me_checkout_shippingAddress_country;
   countryArea: string;
   phone: string | null;
-  /**
-   * Address is user's default billing address.
-   */
-  isDefaultBillingAddress: boolean | null;
-  /**
-   * Address is user's default shipping address.
-   */
-  isDefaultShippingAddress: boolean | null;
 }
 
 export interface UserCheckoutDetails_me_checkout_availableShippingMethods_price {
@@ -518,6 +497,18 @@ export interface UserCheckoutDetails_me_checkout_lines {
   variant: UserCheckoutDetails_me_checkout_lines_variant;
 }
 
+export interface UserCheckoutDetails_me_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface UserCheckoutDetails_me_checkout {
   __typename: "Checkout";
   /**
@@ -529,7 +520,6 @@ export interface UserCheckoutDetails_me_checkout {
    * The ID of the object.
    */
   id: string;
-  user: UserCheckoutDetails_me_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -557,6 +547,14 @@ export interface UserCheckoutDetails_me_checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (UserCheckoutDetails_me_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: UserCheckoutDetails_me_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface UserCheckoutDetails_me {