+++ src/checkout/types/Checkout.ts
@@ -30,11 +30,6 @@ export interface Checkout_availablePaymentGateways {
   config: Checkout_availablePaymentGateways_config[];
 }
 
-export interface Checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface Checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -502,6 +497,18 @@ export interface Checkout_lines {
   variant: Checkout_lines_variant;
 }
 
+export interface Checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface Checkout {
   __typename: "Checkout";
   /**
@@ -513,7 +520,6 @@ export interface Checkout {
    * The ID of the object.
    */
   id: string;
-  user: Checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -541,4 +547,12 @@ export interface Checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (Checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: Checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }