+++ src/checkout/types/createCheckout.ts
@@ -45,11 +45,6 @@ export interface createCheckout_checkoutCreate_checkout_availablePaymentGateways
   config: createCheckout_checkoutCreate_checkout_availablePaymentGateways_config[];
 }
 
-export interface createCheckout_checkoutCreate_checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface createCheckout_checkoutCreate_checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -517,6 +512,18 @@ export interface createCheckout_checkoutCreate_checkout_lines {
   variant: createCheckout_checkoutCreate_checkout_lines_variant;
 }
 
+export interface createCheckout_checkoutCreate_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface createCheckout_checkoutCreate_checkout {
   __typename: "Checkout";
   /**
@@ -528,7 +535,6 @@ export interface createCheckout_checkoutCreate_checkout {
    * The ID of the object.
    */
   id: string;
-  user: createCheckout_checkoutCreate_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -556,6 +562,14 @@ export interface createCheckout_checkoutCreate_checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (createCheckout_checkoutCreate_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: createCheckout_checkoutCreate_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface createCheckout_checkoutCreate {