+++ src/checkout/types/getCheckout.ts
@@ -30,11 +30,6 @@ export interface getCheckout_checkout_availablePaymentGateways {
   config: getCheckout_checkout_availablePaymentGateways_config[];
 }
 
-export interface getCheckout_checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface getCheckout_checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -502,6 +497,18 @@ export interface getCheckout_checkout_lines {
   variant: getCheckout_checkout_lines_variant;
 }
 
+export interface getCheckout_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface getCheckout_checkout {
   __typename: "Checkout";
   /**
@@ -513,7 +520,6 @@ export interface getCheckout_checkout {
    * The ID of the object.
    */
   id: string;
-  user: getCheckout_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -541,6 +547,14 @@ export interface getCheckout_checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (getCheckout_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: getCheckout_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface getCheckout {