+++ src/checkout/views/Review/View.tsx
@@ -6,12 +6,13 @@ import { AlertManager, useAlert } from "react-alert";
 import { generatePath, RouteComponentProps } from "react-router";
 import { Link } from "react-router-dom";
 
+import { Money } from "@components/containers";
+
 import { Button, CartTable } from "../../../components";
 
 import { CartContext } from "../../../components/CartProvider/context";
 import { extractCheckoutLines } from "../../../components/CartProvider/utils";
 import { orderConfirmationUrl } from "../../../routes";
-import { StepCheck } from "../../components";
 import { CheckoutContext } from "../../context";
 import { paymentUrl } from "../../routes";
 import { TypedCompleteCheckoutMutation } from "./queries";
@@ -50,7 +51,6 @@ const completeCheckout = (
 const View: React.FC<RouteComponentProps<{ token?: string }>> = ({
   history,
   match: {
-    path,
     params: { token },
   },
 }) => {
@@ -60,21 +60,13 @@ const View: React.FC<RouteComponentProps<{ token?: string }>> = ({
     dummyStatus,
     checkout,
     clear: clearCheckout,
-    step,
   } = React.useContext(CheckoutContext);
   const { clear: clearCart } = React.useContext(CartContext);
 
-  const stepCheck = (
-    <StepCheck checkout={checkout} step={step} path={path} token={token} />
-  );
-
-  if (!checkout) {
-    return stepCheck;
-  }
+  const discountExists = checkout.discount && !!checkout.discount.amount;
 
   return (
     <>
-      {stepCheck}
       <div className="checkout-review">
         <Link
           to={generatePath(paymentUrl, { token })}
@@ -84,7 +76,7 @@ const View: React.FC<RouteComponentProps<{ token?: string }>> = ({
         </Link>
 
         <div className="checkout__step checkout__step--inactive">
-          <span>5</span>
+          <span>{checkout.isShippingRequired ? "5" : "3"}</span>
           <h4 className="checkout__header">Review your order</h4>
         </div>
 
@@ -92,14 +84,26 @@ const View: React.FC<RouteComponentProps<{ token?: string }>> = ({
           <CartTable
             lines={extractCheckoutLines(checkout.lines)}
             subtotal={checkout.subtotalPrice.gross.localized}
-            deliveryCost={checkout.shippingMethod.price.localized}
+            deliveryCost={
+              checkout.shippingMethod && checkout.shippingMethod.price.localized
+            }
             totalCost={checkout.totalPrice.gross.localized}
+            discount={
+              discountExists && (
+                <>
+                  - <Money money={checkout.discount} />
+                </>
+              )
+            }
+            discountName={checkout.discountName}
           />
           <div className="checkout-review__content">
             <Summary
               checkout={checkout}
               cardData={cardData}
               dummyStatus={dummyStatus}
+              history={history}
+              token={token}
             />
             <div className="checkout-review__content__submit">
               <TypedCompleteCheckoutMutation