+++ src/checkout/views/Shipping/types/updateCheckoutShippingAddress.ts
@@ -45,11 +45,6 @@ export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_che
   config: updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_availablePaymentGateways_config[];
 }
 
-export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -517,6 +512,18 @@ export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_che
   variant: updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_lines_variant;
 }
 
+export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout {
   __typename: "Checkout";
   /**
@@ -528,7 +535,6 @@ export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_che
    * The ID of the object.
    */
   id: string;
-  user: updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -556,6 +562,14 @@ export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate_che
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: updateCheckoutShippingAddress_checkoutShippingAddressUpdate_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface updateCheckoutShippingAddress_checkoutShippingAddressUpdate {
@@ -592,11 +606,6 @@ export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_avai
    * Payment gateway client configuration.
    */
   config: updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_availablePaymentGateways_config[];
-}
-
-export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_user {
-  __typename: "User";
-  email: string;
 }
 
 export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_totalPrice_gross {
@@ -1066,6 +1075,18 @@ export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_line
   variant: updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_lines_variant;
 }
 
+export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout {
   __typename: "Checkout";
   /**
@@ -1077,7 +1098,6 @@ export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout {
    * The ID of the object.
    */
   id: string;
-  user: updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -1105,6 +1125,14 @@ export interface updateCheckoutShippingAddress_checkoutEmailUpdate_checkout {
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: updateCheckoutShippingAddress_checkoutEmailUpdate_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface updateCheckoutShippingAddress_checkoutEmailUpdate_errors {