+++ src/checkout/views/ShippingOptions/types/updateCheckoutShippingOptions.ts
@@ -43,11 +43,6 @@ export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_chec
   config: updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_availablePaymentGateways_config[];
 }
 
-export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_user {
-  __typename: "User";
-  email: string;
-}
-
 export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_totalPrice_gross {
   __typename: "Money";
   /**
@@ -515,6 +510,18 @@ export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_chec
   variant: updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_lines_variant;
 }
 
+export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_discount {
+  __typename: "Money";
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Amount of money.
+   */
+  amount: number;
+}
+
 export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout {
   __typename: "Checkout";
   /**
@@ -526,7 +533,6 @@ export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_chec
    * The ID of the object.
    */
   id: string;
-  user: updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_user | null;
   /**
    * The sum of the the checkout line prices, with all the taxes,shipping costs, and discounts included.
    */
@@ -554,6 +560,14 @@ export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate_chec
    * A list of checkout lines, each containing information about an item in the checkout.
    */
   lines: (updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_lines | null)[] | null;
+  /**
+   * Returns True, if checkout requires shipping.
+   */
+  isShippingRequired: boolean;
+  discount: updateCheckoutShippingOptions_checkoutShippingMethodUpdate_checkout_discount | null;
+  discountName: string | null;
+  translatedDiscountName: string | null;
+  voucherCode: string | null;
 }
 
 export interface updateCheckoutShippingOptions_checkoutShippingMethodUpdate {