+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Change register mutation to accountRegister - #549 by @tomaszszymanski129
 - Add `ProductVariantPicker` component supporting multiple product variant attributes - #550 by @orzechdev
 - Fix not working storefront when no data in saleor database exist - #551 by @orzechdev
+- Make checkout working without shipping if it is not required - #571 by @orzechdev
+- Add ability to apply a promo code in checkout - #582 by @orzechdev
 
 ## 0.7.0